fix(build): split TRMNL types into a server-free module
\`lib/trmnl/registry.ts\` does top-level \`fs.readFile\`/network IO. Even
\`import type { TrmnlModel }\` from a \"use client\" file dragged the
whole module into the client bundle through Turbopack's tree-shaker,
which doesn't fully erase types from a module with side-effecting
imports — \`pnpm build\` failed with \"node:fs/promises is an external
module\".

Move the pure-data definitions (\`TrmnlModel\`, \`TrmnlPalette\`,
\`RegistryResource\`, \`DeviceProfile\`, \`DEFAULT_MODEL_NAME\`) into
\`lib/trmnl/types.ts\`. \`registry.ts\` and \`device-profile.ts\`
re-export from there for back-compat. Client and server modules now
import types from \`@/lib/trmnl/types\`, leaving the registry runtime
purely server-side.
