+++ app/(app)/device/[friendly_id]/client-page.tsx
@@ -15,7 +15,7 @@ import {
 	DEFAULT_IMAGE_HEIGHT,
 	DEFAULT_IMAGE_WIDTH,
 } from "@/lib/recipes/constants";
-import type { TrmnlModel, TrmnlPalette } from "@/lib/trmnl/registry";
+import type { TrmnlModel, TrmnlPalette } from "@/lib/trmnl/types";
 import type { Device, Mixup, Playlist, PlaylistItem } from "@/lib/types";
 import {
 	generateApiKey,
