+++ components/device/device-view.tsx
@@ -14,7 +14,7 @@ import {
 	DEFAULT_IMAGE_HEIGHT,
 	DEFAULT_IMAGE_WIDTH,
 } from "@/lib/recipes/constants";
-import type { TrmnlModel, TrmnlPalette } from "@/lib/trmnl/registry";
+import type { TrmnlModel, TrmnlPalette } from "@/lib/trmnl/types";
 import type { Device } from "@/lib/types";
 import { cn } from "@/lib/utils";
 import {
