+++ lib/database/sql-statements.ts
@@ -951,7 +951,39 @@ GRANT SELECT, INSERT, UPDATE, DELETE ON mixup_slots TO byos_app;`,
 );
 
 CREATE INDEX IF NOT EXISTS idx_plugin_settings_user_id ON plugin_settings(user_id);
-CREATE INDEX IF NOT EXISTS idx_plugin_settings_plugin_id ON plugin_settings(plugin_id);`,
+CREATE INDEX IF NOT EXISTS idx_plugin_settings_plugin_id ON plugin_settings(plugin_id);
+
+-- =============================================================================
+-- RLS: tenant ownership enforced at the DB layer (matches 0014_harden_rls.sql)
+-- =============================================================================
+
+ALTER TABLE plugin_settings ENABLE ROW LEVEL SECURITY;
+ALTER TABLE plugin_settings FORCE ROW LEVEL SECURITY;
+
+DROP POLICY IF EXISTS plugin_settings_select_policy ON plugin_settings;
+DROP POLICY IF EXISTS plugin_settings_insert_policy ON plugin_settings;
+DROP POLICY IF EXISTS plugin_settings_update_policy ON plugin_settings;
+DROP POLICY IF EXISTS plugin_settings_delete_policy ON plugin_settings;
+
+CREATE POLICY plugin_settings_select_policy ON plugin_settings
+    FOR SELECT
+    USING (user_id = current_setting('app.current_user_id', true));
+
+CREATE POLICY plugin_settings_insert_policy ON plugin_settings
+    FOR INSERT
+    WITH CHECK (user_id = current_setting('app.current_user_id', true));
+
+CREATE POLICY plugin_settings_update_policy ON plugin_settings
+    FOR UPDATE
+    USING (user_id = current_setting('app.current_user_id', true))
+    WITH CHECK (user_id = current_setting('app.current_user_id', true));
+
+CREATE POLICY plugin_settings_delete_policy ON plugin_settings
+    FOR DELETE
+    USING (user_id = current_setting('app.current_user_id', true));
+
+GRANT SELECT, INSERT, UPDATE, DELETE ON plugin_settings TO byos_app;
+GRANT USAGE, SELECT ON SEQUENCE plugin_settings_id_seq TO byos_app;`,
 	},
 	validate_schema: {
 		title: "Validate Database Schema",
